Add G_PureFile to gametype AS
This allows gametypes to register files other than sounds and models as
pure. This can be used to transfer other files to connecting clients,
for example ui files.

AS can't just call G_ModelIndex with pure true because the function
checks the returned modelindex before registering is as pure. However we
piggyback G_PureModel as it's just passed through to the backend without
checks. This way we can avoid an extension of the import/export
interfaces.