Adds some administration functions to AS
This changeset allows gametypes to
* promote/demote operators
* mute/unmute clients
* drop clients with custom messages printed on the clients' screen after
dropping out of the game.

This functionality allows gametypes to implement an admin interface to
server admins. While I ultimately believe that this interface should be
included into the game itself and should not be a concern of the
gametype, I believe managing this behaviour from AS is good enough for
right now until that is done.