Fix vsays using the wrong sounds when dead.
This works around game setting ent->s.modelindex to 0 in various places. Game also
depends on modelindex == 0 for G_ISGHOSTING among others. Instead of changing all these
interactions, this workaround synchronizes the chosen playermodel for each player via configstrings
in addition to the modelindex. If the modelindex received from the server is not valid, cgame
first falls back to the model provided using configstrings and only if that is invalid as well
cgame uses DEFAULT_PLAYERMODEL.