Fixed an oversight causing long playernames to overflow configstrings.
The addition of models into configstrings did not account for
MAX_CONFIGSTRING_CHARS. Therefore a long player name with a lot of
colors could cause the configstring to be an invalid InfoString. This
caused players to be kicked.

This quick and dirty fix works around this problem by only adding the model
to the playerinfo configstring if there's enough space available. If there's
no modelindex in the playerinfo, cgame falls back to DEFAULT_PLAYERMODEL.

This workaround should be replaced by a proper fix.