Remove Warshell for headhunt matches
We use the Warshell effect (EF_SHELL) on the tagged player in headhunt, so this will reduce confusion. Also, the scoreboard detects who's carrying Warshell by the effect, so this also ensures that the scoreboard will always only show the tagged player with the Warshell icon during headhunt, rather than possibly showing two players with the icon.